renames the files in ContainerSelect
