--- src/signals/incident/components/form/ContainerSelect/ContainerLayer/index.ts
@@ -1,3 +1,0 @@
-import ContainerLayer from './ContainerLayer';
-
-export default ContainerLayer;
