--- src/signals/incident/components/form/ContainerSelect/ContainerList/index.js
@@ -1,3 +1,0 @@
-import ContainerList from './ContainerList';
-
-export default ContainerList;
