+++ src/signals/incident/components/form/ContainerSelect/LegendPanel/index.test.tsx
@@ -2,8 +2,8 @@ import { render, screen } from '@testing-library/react';
 import React from 'react';
 import { withAppContext } from 'test/utils';
 
-import LegendPanel from './LegendPanel';
-import type { LegendPanelProps } from './LegendPanel';
+import LegendPanel from '.';
+import type { LegendPanelProps } from '.';
 
 describe('LegendPanel', () => {
   const props: LegendPanelProps = {
