--- src/signals/incident/components/form/ContainerSelect/LegendPanel/index.ts
@@ -1,3 +1,0 @@
-import LegendPanel from './LegendPanel';
-
-export default LegendPanel;
