+++ src/signals/incident/components/form/ContainerSelect/LegendToggleButton/index.test.tsx
@@ -5,8 +5,8 @@ import { render, screen } from '@testing-library/react';
 
 import { withAppContext } from 'test/utils';
 
-import LegendToggleButton from './LegendToggleButton';
-import type { LegendToggleButtonProps } from './LegendToggleButton';
+import LegendToggleButton from '.';
+import type { LegendToggleButtonProps } from '.';
 import { MapPanelProvider } from '@amsterdam/arm-core';
 import { SnapPoint } from '@amsterdam/arm-core/lib/components/MapPanel/constants';
