--- src/signals/incident/components/form/ContainerSelect/Selector/index.ts
@@ -1,3 +1,0 @@
-import Selector from './Selector';
-
-export default Selector;
