--- src/signals/incident/components/form/ContainerSelect/ViewerContainer/index.ts
@@ -1,3 +1,0 @@
-import ViewerContainer from './ViewerContainer';
-
-export default ViewerContainer;
