+++ src/signals/incident/components/form/ContainerSelect/WfsLayer/index.test.tsx
@@ -10,8 +10,8 @@ import { Map } from '@amsterdam/react-maps';
 import containersJson from 'utils/__tests__/fixtures/containers.json';
 import MAP_OPTIONS from 'shared/services/configuration/map-options';
 import type { DataLayerProps } from '../types';
-import WfsDataContext, { NO_DATA } from './WfsDataContext';
-import WfsLayer, { isLayerVisible } from './WfsLayer';
+import WfsDataContext, { NO_DATA } from './context';
+import WfsLayer, { isLayerVisible } from '.';
 
 const fetchMock = fetch as FetchMock;
