--- src/signals/incident/components/form/ContainerSelect/WfsLayer/index.ts
@@ -1,3 +1,0 @@
-import WfsLayer from './WfsLayer';
-
-export default WfsLayer;
