+++ src/signals/incident/components/form/ContainerSelect/index.test.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import incidentJson from 'utils/__tests__/fixtures/incident.json';
 import { withAppContext } from 'test/utils';
-import type { ContainerSelectProps } from './ContainerSelect';
-import ContainerSelect from './ContainerSelect';
-import { initialValue } from './ContainerSelectContext';
-import { withContainerSelectContext } from './ContainerSelectContext.test';
+import type { ContainerSelectProps } from '.';
+import ContainerSelect from '.';
+import { initialValue } from './context';
+import { withContainerSelectContext } from './context.test';
 import type { Location } from 'types/incident';
 import userEvent from '@testing-library/user-event';
