--- src/signals/incident/components/form/ContainerSelect/index.ts
@@ -1,3 +1,0 @@
-import ContainerSelect from './ContainerSelect';
-
-export default ContainerSelect;
